Migrate to Dalamud API 15 (Dalamud.NET.Sdk/15.0.0), bump version to 1.5.0
- Replace Microsoft.NET.Sdk with Dalamud.NET.Sdk/15.0.0, removing explicit DLL references
- Update IChatGui.ChatMessage handlers to new IHandleableChatMessage signature
- Fix TerritoryType/TerritoryChanged uint→ushort casts (API 15 type change)
- Remove out-of-range XivChatType cast (>110 now invalid in API 15)
- Update DalamudApiLevel to 15 in Flowline.json and repo.json

+++ Flowline/Services/DutyDetectionService.cs
@@ -45,9 +45,9 @@ public class DutyDetectionService : IDisposable
         CheckCurrentTerritory();
     }
 
-    private void OnTerritoryChanged(ushort territoryId)
+    private void OnTerritoryChanged(uint territoryId)
     {
-        currentTerritoryId = territoryId;
+        currentTerritoryId = (ushort)territoryId;
         CheckCurrentTerritory();
     }
